fix branch of test to actually apply
the if test was pointless, as a regexp object would not equal the empty
string. I would be a regexp object with `/(?:)/` as the regex. Instead,
change the conditional to test the variable that holds the string before
being put into the regexp object.

the test now avoids some pointless regex matching when allow/deny regex are
empty.